Created TestBooleanAlgebra.java
